Messages
Added message to the pile of bones directly to the left of the door leading to the boss room.
Added message to the locked door in the boss room that changes depending upon whether the player has the key in their inventory or not.
Boss now drops the key.
Fixed bug with blue projectiles not slowing the player down.
Fixed bug where cultist projectiles were at weird angles.