Added Test Room And Various Bug Fixes
Added the test room with all object types (excluding monsters), as well as the option to go there from the title menu. Changed how the obj_Climb object moved the player (room and coordinates must now be specified in the instance creation code). Changed the warp object to freeze the player temporarily when entering a new room to prevent them from clipping into another object after the new room has been loaded. Fixed a water tile that was in the wrong layer in room0.