Debuff Fix
If the debuff object is destroyed while an alarm is still set then it will run the alarm's code prematurely. This means that the player will be set back to normal when they die, leave a room, or do anything else to destroy the debuff before it is supposed to.